# apps/wallet-web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "relational-rust-server",
    "description": "Custodial Avalanche wallet service running inside Intel SGX enclave with Gramine.",
    "contact": {
      "name": "Relational Network"
    },
    "license": {
      "name": "AGPL-3.0-or-later",
      "identifier": "AGPL-3.0-or-later"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/wallets": {
      "get": {
        "tags": [
          "Wallets"
        ],
        "summary": "List all wallets owned by the authenticated user.",
        "operationId": "list_wallets",
        "responses": {
          "200": {
            "description": "List of wallets",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Wallets"
        ],
        "summary": "Create a new wallet for the authenticated user.",
        "description": "Generates a new p256 keypair inside the SGX enclave and stores it\nencrypted on disk. Returns the wallet metadata (never the private key).",
        "operationId": "create_wallet",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWalletRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Wallet created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateWalletResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/v1/wallets/{wallet_id}": {
      "get": {
        "tags": [
          "Wallets"
        ],
        "summary": "Get a specific wallet by ID.",
        "description": "Only returns wallets owned by the authenticated user.",
        "operationId": "get_wallet",
        "parameters": [
          {
            "name": "wallet_id",
            "in": "path",
            "description": "Wallet ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Wallet details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - not your wallet"
          },
          "404": {
            "description": "Wallet not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Wallets"
        ],
        "summary": "Delete (soft-delete) a wallet.",
        "description": "Marks the wallet as deleted. The private key is retained for potential\nrecovery but the wallet cannot be used for new transactions.",
        "operationId": "delete_wallet",
        "parameters": [
          {
            "name": "wallet_id",
            "in": "path",
            "description": "Wallet ID to delete",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Wallet deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteWalletResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - not your wallet"
          },
          "404": {
            "description": "Wallet not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateWalletRequest": {
        "type": "object",
        "description": "Request to create a new wallet.",
        "properties": {
          "label": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional human-readable label for the wallet."
          }
        }
      },
      "CreateWalletResponse": {
        "type": "object",
        "description": "Response after creating a wallet.",
        "required": [
          "wallet",
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "description": "Message indicating success."
          },
          "wallet": {
            "$ref": "#/components/schemas/WalletResponse",
            "description": "The created wallet details."
          }
        }
      },
      "DeleteWalletResponse": {
        "type": "object",
        "description": "Response after deleting a wallet.",
        "required": [
          "message",
          "wallet_id"
        ],
        "properties": {
          "message": {
            "type": "string",
            "description": "Message indicating success."
          },
          "wallet_id": {
            "type": "string",
            "description": "The ID of the deleted wallet."
          }
        }
      },
...
      "WalletListResponse": {
        "type": "object",
        "description": "Response containing a list of wallets.",
        "required": [
          "wallets",
          "total"
        ],
        "properties": {
          "total": {
            "type": "integer",
            "description": "Total count of wallets.",
            "minimum": 0
          },
          "wallets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WalletResponse"
            },
            "description": "List of wallets owned by the user."
          }
        }
      },
      "WalletResponse": {
        "type": "object",
        "description": "Response returned to API clients (never includes private key).",
        "required": [
          "wallet_id",
          "public_address",
          "created_at",
          "status"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the wallet was created"
          },
          "label": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional human-readable label"
          },
          "public_address": {
            "type": "string",
            "description": "Public address"
          },
          "status": {
            "$ref": "#/components/schemas/WalletStatus",
            "description": "Current wallet status"
          },
          "wallet_id": {
            "type": "string",
            "description": "Unique wallet identifier"
          }
        }
      },
      "WalletStatus": {
        "type": "string",
        "description": "Wallet status.",
        "enum": [
          "active",
          "suspended",
          "deleted"
        ]
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Clerk JWT token"
      }
    }
  },
...
}